requirements.txt
flask==3.1.0
gunicorn==23.0.0
openai==1.58.1
python-dotenv==1.0.1
supabase==2.13.0
